Prevent overwriting release tags
When building docker images locally, it is possible to accidentally
overwrite a released tag due to improper configuration. This isn't a
good idea because it can cause a lot of confusion when the locally
tagged image is no longer the same as the image that would be pulled
down by from the registry. Prevent this by checking if the tag looks
like a release tag, and erroring out with a helpful messages instead of
overwriting it.